+++ Changelog.md
@@ -5,6 +5,7 @@
 ### Bugfixes
 * (De)serializing of blank node Entities with Constrains did not keep the proper link
 * Allow to configure certain settings via `.iadopt-catalogue.yml`
+* cope with Variables that have no title - using local part of IRI if needed
 
 ## 0.1.0